Fix stability issues using double precision for sums (slower but determenistic)

+++ KMeansProject/KMeansProject/GPUSolution/k_means_data.cu
@@ -5,7 +5,7 @@ void KMeansData::allocate_memory()
 	CHECK_CUDA(cudaMalloc((void**)&d_points, n * d * sizeof(float)));
 	CHECK_CUDA(cudaMalloc((void**)&d_centroids, k * d * sizeof(float)));
 	CHECK_CUDA(cudaMalloc((void**)&d_labels, n * sizeof(int)));
-	CHECK_CUDA(cudaMalloc((void**)&d_sums, k * d * sizeof(float)));
+	CHECK_CUDA(cudaMalloc((void**)&d_sums, k * d * sizeof(double)));
 	CHECK_CUDA(cudaMalloc((void**)&d_counts, k * sizeof(int)));
 	CHECK_CUDA(cudaMalloc((void**)&d_changes_count, sizeof(int)));
 }
@@ -45,6 +45,6 @@ void KMeansData::fill_gpu_data(const std::vector<float>& h_points_soa, const std
 	CHECK_CUDA(cudaMemcpy(d_centroids, h_centroids_soa.data(), k * d * sizeof(float), cudaMemcpyHostToDevice));
 	CHECK_CUDA(cudaMemset(d_labels, 0, n * sizeof(int)));
 	CHECK_CUDA(cudaMemset(d_counts, 0, k * sizeof(int)));
-	CHECK_CUDA(cudaMemset(d_sums, 0.0F, k * d * sizeof(float)));
+	CHECK_CUDA(cudaMemset(d_sums, 0, k * d * sizeof(double)));
 	CHECK_CUDA(cudaMemset(d_changes_count, 0, sizeof(int)));
 }
